requirements.txt
gradio==4.19.2
openai>=1.30.0
python-dotenv==1.0.0